% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gtfs_trips.R
\name{find_gtfs_trips}
\alias{find_gtfs_trips}
\title{Find gtfs trips}
\usage{
find_gtfs_trips(routes, gtfs_path)
}
\arguments{
\item{routes}{A data.table with col 'route_id', the variable to
filter the trips.}

\item{gtfs_path}{A character vector with the path to the GTFS files.}
}
\value{
A data.table with all the trips found
}
\description{
\code{find_gtfs_trips()} takes as input a data.table with col 'route_id' with the
routes that got a matched along with col 'gtfs_file' with the date of the
files to look for. It then filters the GTFS stop_times file to load the trips
for those routes. Finally, it filters those trips to extract the longest
trip for each route and returns a data.table of trips.
}
