% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gtfs_trips.R
\name{fix_gtfs_trips}
\alias{fix_gtfs_trips}
\title{Removes the first or last point that creates a loop in gtfs trips}
\usage{
fix_gtfs_trips(trips)
}
\arguments{
\item{trips}{A data.table returned by \code{get_gtfs_stops()}.}
}
\description{
Some gtfs trips have errors with duplicated stops, for some reason not yet
known, the last stop of the trip is linked to the first stop creating a loop
for some trips. \code{fix_gtfs_trips()} removes the points that creates those
loops.
}
