% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cenefa}
\alias{get_cenefa}
\title{Get Cenefa}
\usage{
get_cenefa(parada_vector)
}
\arguments{
\item{parada_vector}{A chracter vector with unique bus stop names}
}
\value{
A character vector with clean stop codes
}
\description{
Utility function to extract clean stop codes from validation file column
parada.
}
