% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_stops.R
\name{get_gtfs_stops}
\alias{get_gtfs_stops}
\title{Get gtfs stops data for trips}
\usage{
get_gtfs_stops(trips, gtfs_path)
}
\arguments{
\item{trips}{A data.table with all the trips found.}

\item{gtfs_path}{A character vector with the path to the GTFS files.}
}
\value{
A data.table with all the trips found and all data for the stops in
those trips.
}
\description{
\code{get_gtfs_stops()} Joins data.table trips with GTFS stops, route and trips
to get all the trip's data in one data.table
}
