% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_short_name}
\alias{get_short_name}
\title{Extract a Clean Route Name From Validation File's Routes Column}
\usage{
get_short_name(raw_rutas)
}
\arguments{
\item{raw_rutas}{A character vector with unique route's name from
validation file.}
}
\value{
A data.table with clean route names
}
\description{
Extract a Clean Route Name From Validation File's Routes Column
}
