% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_stops_by_route.R
\name{get_unique_stops_by_route}
\alias{get_unique_stops_by_route}
\title{Get unique stops by route}
\usage{
get_unique_stops_by_route(raw_dt)
}
\arguments{
\item{raw_data}{A data.table with route's data from the validation file.}
}
\value{
A list of data.tables, a data.table for each route.
}
\description{
\code{get_unique_stops_by_route()} returns a list of data.tables with each route
in the validation file as an element.
}
\section{Column names for each data.tables in the list}{

\itemize{
\item ruta: Validation file route name.
\item parada: Validation file stop name.
\item route_short_name: Clean stop name to match against the GTFS file, double
letter prefix are not split yet.
\item stop_code: Clean stop code to match against the GTFS file.
}
}

