% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_unzip.R
\name{gtfs_to_unzip}
\alias{gtfs_to_unzip}
\title{Select the gtfs files by date}
\usage{
gtfs_to_unzip(date, gtfs_path)
}
\arguments{
\item{date}{A character string in the 'YYYY-MM-DD' format with the date of the
validation file to process.}

\item{gtfs_path}{A character string with the path to the gtfs files.}
}
\value{
A character vector with GTFS files selected by dates.
}
\description{
\code{gtfs_to_unzip} returns a character vector with GTFS files selected by dates.
}
