% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_val_routes.R
\name{load_val_routes}
\alias{load_val_routes}
\title{Load Route's Data From Validation File}
\usage{
load_val_routes(date, val_file_path)
}
\arguments{
\item{date}{A character string with a date in the format 'YYYY-MM-DD' to
match a date written on the validation file name.}

\item{val_file_path}{A character string with the path to the directory where
the validation files are stored.}
}
\value{
A data.table with colums needed to process all routes in
the validation file.
}
\description{
\code{load_val_routes()} reads a validation file selected by \code{date}. It takes as
input a date that matches the validation file name and a path to the
directory where the validation files are stored. It then reads the selected
file and returns a data.table with the variables relevant to process the
routes
}
