% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ruta_to_route_id.R
\name{map_ruta_to_route_id}
\alias{map_ruta_to_route_id}
\title{Map routes from validation files to gtfs routes}
\usage{
map_ruta_to_route_id(
  date,
  val_file_path = "../data_sets/sitp/validaciones_TM/dia",
  gtfs_path = "~/R/data_sets/sitp/gtfs/"
)
}
\arguments{
\item{date}{A character string in the 'YYYY-MM-DD' format with the date of the
validation file to process, this date is found on the name of the file. Mind
that the date on the file name is in a 'YYYYMMDD' format with no dashes, do
not use that format, it will run into an error.}

\item{val_file_path}{A character string with the path to the input validation files.}

\item{gtfs_path}{A character string with the path to the gtfs files.}
}
\value{
\code{map_ruta_to_route_id()} returns a data.table with clean names for
routes and bus stops with all data needed to create a GTFS feed from the
original validation file.
}
\description{
\code{map_ruta_to_route_id()} extracts all routes from the validation file and
maps them to the routes on the gtfs file based on the bus stops along the
routes.
}
\section{Files needed to run the function}{


This function process two different kind of data files, a
validation file and a GTFS file. Both files are available for download at:
\itemize{
\item Validation files at:
'https://storage.googleapis.com/validaciones_tmsa/ValidacionZonal.html'
\item GTFS files at:
'https://datosabiertos-transmilenio.hub.arcgis.com/search?groupIds=ca6e3d0acf57461d91228659c1b0d2dd'
}

The gtfs files have the geographical data we need, also, this function
only requires gtfs files from dates as old as the oldest validation
file available. No need to download all of them.

Do not modify in any way the names of the downloaded files, the function relies
on them.

The validation file contains all data from the smart travel cards that users
validate when boarding the buses. It has many variables but in this function
we only use a few:
\itemize{
\item 'Rutas' - The name of the bus route
\item 'Estacion_Parada' - Bus stop code where the user got on the bus
\item 'Fecha_Transaccion' - A time stamp at boarding time
\item 'ID_Vehiculo' - Id number of the bus taken
}
}

\section{The problem to solve}{


With these four variables is possible to recreate the routes for each bus in
the system. We are only missing the geographical coordinates of the bus stops
to create an accurate bus trip. This is where the GTFS file comes in.
The validation file does not provide geo coordinates, only bus stops codes,
on the other hand, the GTFS file does have route information and geo coordinates
for bus stops. The main problem to map the routes between these two files
is that the route names they use are different in each file. This problem
can be solve via bus stop code names which are common on both files.

\code{map_ruta_to_route_id()} solves the above mentioned problem by mapping the
bus stops from the validation file to the bus stops on the GTFS file allowing
to map the routes in both files.
}

