% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_stops_to_route_id.R
\name{map_stops_to_route_id}
\alias{map_stops_to_route_id}
\title{Map stops from validation files to gtfs routes}
\usage{
map_stops_to_route_id(date, val_file_path, gtfs_path)
}
\arguments{
\item{date}{A character string in the 'YYYY-MM-DD' format with the date of the
validation file to process, this date is found on the name of the file. Mind
that the date on the file name is in a 'YYYYMMDD' format with no dashes, do
not use that format, it will run into an error.}

\item{val_file_path}{A character string with the path to the input validation files.}

\item{gtfs_path}{A character string with the path to the gtfs files.}
}
\value{
\code{map_ruta_to_route_id()} returns a data.table with clean names for
routes and bus stops with all data needed to recreate a GTFS feed from the
original validation file.
}
\description{
\code{map_stops_to_route_id()} extracts the routes data from the
validation file (read the details section for the extracted variables) to
recreate the trips made by the buses on the different routes by day. With
this trips table it then selects and extracts the requiered data from the
gtfs files to map the routes on the validation file to the routes on the
gtfs files, it then returns a data.table with the mapped routes based on
the bus stops matched in both files.
}
\section{Files needed to run the function}{


This function process two different kind of data files, a
validation file and a GTFS file. Both files are available for download at:
\itemize{
\item Validation files at:
'https://storage.googleapis.com/validaciones_tmsa/ValidacionZonal.html'
\item GTFS files at:
'https://datosabiertos-transmilenio.hub.arcgis.com/search?groupIds=ca6e3d0acf57461d91228659c1b0d2dd'
}

The validation file contains all data from the smart travel cards that users
validate when boarding the buses. It has many variables but in this function
we only use a few:
\itemize{
\item 'Rutas' - The name of the bus route
\item 'Estacion_Parada' - Bus stop code where the user got on the bus
\item 'Fecha_Transaccion' - A time stamp at boarding time
\item 'ID_Vehiculo' - Id number of the bus taken
}

The gtfs files have the geographical data we need, also, this function
only requires gtfs files from dates as old as the oldest validation
file available. No need to download all of them.

Do not modify in any way the names of the downloaded files, the function relies
on them as they are.
}

\section{How to map the routes in both files}{


The main problem to map the routes between these two files is that
the route names are different in each file. This is solved via bus stop code
names which are common on both files.

The validation file has bus codes but lacks their geographical coordinates,
the function takes those from the GTFS files.

\code{map_ruta_to_route_id()} solves the above mentioned problem by mapping the
bus stops from the validation file to the bus stops on the GTFS file allowing
to map the routes in both files.
}

