% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_routes_in_gtfs.R
\name{match_routes_in_gtfs}
\alias{match_routes_in_gtfs}
\title{Match Routes in GTFStf}
\usage{
match_routes_in_gtfs(rutas, gtfs)
}
\arguments{
\item{rutas}{A list of data.tables, one data.table for each route.}

\item{gtfs}{A list of gtfs data.tables with route names to match to.}
}
\description{
\code{match_routes_in_gtfs()} matches route names from the validation files to
route ids in the GTFS files. It returns a data.table with all route ids
that match to a route name in the validation file.
}
\details{
\code{match_routes_in_gtfs()} takes route names from the validation file
and finds matches on gtfs files sorted by date. First it tries matches
on the most recent gtfs file and stores this matches. Failed route matches
are then processed further by spliting the double letter prefix from their
name and  are matched again against the latest gtfs file and stored . Matches that
fail this second round are then tried against the older gtfs files until they
get a match in any of them. The trial goes from the most recent gtfs file down
to the oldest to avoid matching duplicates, at the end of the loop, if there
are routes that did not find a match, they get discarded throwing a warning.
}
