% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gtfs_trips.R
\name{remove_head_tail}
\alias{remove_head_tail}
\title{Remove the first or last point in route on condition}
\usage{
remove_head_tail(sfc_route)
}
\arguments{
\item{sfc_route}{An sf sfc route object.}
}
\value{
The input object with points removed if condition met.
}
\description{
Remove the first or last point in route on condition
}
