% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_web_scrap.R
\name{scrap_routes}
\alias{scrap_routes}
\title{Scrap routes' data from Transmilenio website}
\usage{
scrap_routes()
}
\value{
A data.table
}
\description{
Download routes' codes, names and ids  and return a data.table with those
variables
}
