% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitp.gtfs.R
\docType{package}
\name{sitp.gtfs}
\alias{sitp.gtfs}
\alias{sitp.gtfs-package}
\title{sitp.gtfs: Herramientas para el análisis de frecuencias de las rutas del
sistema de buses públicos de Bogotá SITP}
\description{
Funciones útiles para alaizar las frecuencias de las rutas del SITP a partir
de los archivos de datos de validaciones de usuarios publicados por la
empresa Transmilenio.
}
\author{
\strong{Maintainer}: Federico Viviescas Ramírez \email{federicoviviescas@hotmail.com}

}
\keyword{internal}
