% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_prefix}
\alias{split_prefix}
\title{Split Prefix for Double Letter Route's name}
\usage{
split_prefix(rutas)
}
\arguments{
\item{rutas}{A character vector with unique names for routes that have a double
letter prefix on its name.}
}
\value{
A character vector with unique names with single letter prefix
}
\description{
Split the route's name prefix for route names with double letter name, this
routes missed a match on the first try because the gtfs routes names don't
use double letter prefix.
}
