% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_scale_bar_custom}
\alias{st_scale_bar_custom}
\title{Make scale bar for plots}
\usage{
st_scale_bar_custom(sf_object, pos = "bottomleft", size_km = 1)
}
\arguments{
\item{sfc_object}{An sf sfc object.}
}
\description{
Make scale bar for plots
}
