% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_files_by_date}
\alias{find_files_by_date}
\title{Find validation files by date}
\usage{
find_files_by_date(
  from,
  to = NULL,
  day_num = NULL,
  exclude_holidays = FALSE,
  val_file_path
)
}
\arguments{
\item{from}{A character vector in the format 'YYYY-MM-DD'.}

\item{to}{A character vector in the format 'YYYY-MM-DD'.}

\item{day_num}{A single digit from 0 to 7 where: 0 = search holidays only,
1 = Sunday, 2 = Monday ... and 7 = Saturday. Dates are filtered by day of
the week or if day_num = 0, search for holidays inside the range.}

\item{exclude_holidays}{A logical, 'TRUE' exclude holidays from the search.}

\item{val_file_path}{A character vector with the path to the directory for
the validation files.}
}
\value{
A character vector with vallidation file's names.
}
\description{
\code{find_files_by_date()} returns a character vector of file names that match the
input 'date'.
}
\details{
This function finds data files based on date ranges created from user input
or a single file if argument 'to' is NULL.

These ranges can be filtered by day of the week and by holidays.
The function returns a vector of file names for available validation files
found in given date range, or a single file if no range is given.
}
