% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_gtfs_shapes.R
\name{get_clean_gtfs_shapes}
\alias{get_clean_gtfs_shapes}
\title{Get get_clean_gtfs_shapes}
\usage{
get_clean_gtfs_shapes(route_id, trips)
}
\value{
A gtfs \code{shapes.txt} file.
}
\description{
Using the gtfs \code{stops.txt} file as input, this function returns
a new gtfs \code{shapes.txt} fixing the errors of the original \code{shapes.txt} file.
}
