% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_gtfs_by_date}
\alias{read_gtfs_by_date}
\title{Read GTFS by Date}
\usage{
read_gtfs_by_date(gtfs_zip, gtfs_files)
}
\arguments{
\item{gtfs_zip}{list of gtfs file paths}

\item{gtfs_files}{Character vector with file names to load, e.g. "routes" or "stop_times"}
}
\value{
A list of GTFS data.tables
}
\description{
Load the gtfs files to match routes against them.
}
