% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitp.gtfs.R
\docType{package}
\name{sitp.gtfs}
\alias{sitp.gtfs}
\alias{sitp.gtfs-package}
\title{sitp.gtfs: Tools for frequency analysis of the routes in Bogotá's public bus
system SITP}
\description{
Useful functions for analyzing SITP route frequencies from the user
validation data files published by Transmilenio.
}
\author{
\strong{Maintainer}: Federico Viviescas Ramírez \email{federicoviviescas@hotmail.com}

}
\keyword{internal}
